VITIS-6990:xrt::xclbin API to obtain interface_uuid
